NutriScan-app/requirements.txt
Flask==3.1.0
Pillow==11.1.0
pytesseract==0.3.13
pymongo==4.15.1
gunicorn==23.0.0